Refactored Struct and Addressed Comments
