+++ pkg/models/capturetraffic.go
@@ -1,0 +1,17 @@
+package models
+
+import "time"
+
+type TrafficCaptureParams struct {
+	Path             string
+	ProxyPort        uint32
+	AppCmd           string
+	AppContainer     string
+	AppNetwork       string
+	Delay            uint64
+	BuildDelay       time.Duration
+	Ports            []uint
+	Filters          *TestFilter
+	EnableTele       bool
+	PassThroughHosts []Filters
+}
